+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Changelog
 
 # Todo
-- jsbaton - migrate function dbExec() to new unified jsbaton
 - jenks - replace class-based api with static-function
 - wasm - replace sqljs-api with sqlmath-api in wasm
 - sqlmath - add sql-extension carrayblob()
 - none
 
 # v2022.6.1-beta
+- jsbaton - migrate function dbExec() to new unified jsbaton
 - error - add sqlite errcode SQLITE_ERROR_ZSQL_NULL
 - jsbaton - migrate function dbTableInsert() to new unified jsbaton
 - jsbaton - migrate functions dbClose(), dbMemoryLoadOrSave(), dbNoop(), dbOpen() to new, streamlined jsbaton2